Updating phantom version to 2.1.3
