+++ src/main/scala/org/zouzias/cassandra/example/user/Users.scala
@@ -1,11 +1,24 @@
 package org.zouzias.cassandra.example.user
 
 import java.util.UUID
-import com.websudos.phantom.dsl._
+
+import com.outworkers.phantom.dsl._
+import org.joda.time.DateTime
+
 import scala.concurrent.Future
 
-// The root connector comes from import com.websudos.phantom.dsl._
-abstract class ConcreteUsers extends Users with RootConnector {
+case class User(
+  id: UUID,
+  email: String,
+  name: String,
+  registrationDate: DateTime
+)
+
+class Users extends CassandraTable[Users, User] {
+  object id extends UUIDColumn(this) with PartitionKey
+  object email extends StringColumn(this)
+  object name extends StringColumn(this)
+  object registrationDate extends DateTimeColumn(this)
 
   def store(user: User): Future[ResultSet] = {
     insert.value(_.id, user.id).value(_.email, user.email)
