+++ src/main/scala/org/zouzias/cassandra/example/user/UsersExampleMain.scala
@@ -2,11 +2,10 @@ package org.zouzias.cassandra.example.user
 
 import java.util.UUID
 
-import com.websudos.phantom.connectors.KeySpace
-import com.websudos.phantom.dsl.{ResultSet, DateTime}
+import com.outworkers.phantom.connectors.KeySpace
+import org.joda.time.DateTime
 
-import scala.concurrent.{ExecutionContext, Await}
-import scala.concurrent.duration._
+import scala.concurrent.ExecutionContext
 
 /**
   * Example from http://blog.websudos.com/2015/04/04/a-series-on-phantom-part-1-getting-started-with-phantom/
@@ -15,10 +14,9 @@ object UsersExampleMain extends App{
 
   implicit val ex = ExecutionContext.Implicits.global
   implicit val session = LocalDatabase.session
-  implicit val keySpace = new KeySpace(LocalDatabase.keyspace.name)
+  implicit val keySpace = KeySpace(LocalDatabase.keyspace.name)
 
-  Await.result(LocalDatabase.autocreate.future(), 5.seconds)
-
+  LocalDatabase.create()
 
   val user = User(UUID.randomUUID(), "zouzias@swisscom.com", "Tassos Zouzias", new DateTime())
   LocalDatabase.users.store(user)
