--- src/main/scala/org/zouzias/cassandra/phantom/examples/user/LocalDatabase.scala
@@ -1,14 +1,0 @@
-package org.zouzias.cassandra.example.user
-
-import com.websudos.phantom.connectors.{ContactPoints, KeySpaceDef}
-
-
-object Defaults {
-  val connector = ContactPoints(Seq("192.168.99.100")).keySpace("zouzias")
-}
-
-class LocalDatabase(val keyspace: KeySpaceDef) extends com.websudos.phantom.db.DatabaseImpl(keyspace) {
-  object users extends ConcreteUsers with keyspace.Connector
-}
-
-object LocalDatabase extends LocalDatabase(Defaults.connector)
